Create Pick A Song
Generated a database that stored a set number of songs which included the title, the artist who performed the song, the mood of the song, how long the song was in seconds, and when it was released. I then queried the database with different limitations on what type of song I was looking for.