Create Gradebook
I generated a simple gradebook database that contained the names, grades, and percentage of work completed for each student. For my first query I created a table that displayed the students name, the grade, and a rounded percentage of the total work they completed. For my second query I counted and displayed the number of students who received each of the grade levels.