Create Pokedex Queries
Generated a pokedex containing all of the names, numbers, types, stats, generation, and whether or not they were legendary. The first query shows how many legendary there are vs non legendary. The second query counts the number of each pokemon in each generation. The third query displays the stat average for each stat. The fourth query finds and displays the number, name, and HP of the pokemon with the max possible HP. The fifth query finds and displays all numbers and names of pokemon with Fire as their first type and Flying as their second type. Lastly the sixth query finds and displays the name and defense of each pokemon whose defense is greater than the calculated average defense of all pokemon.