Update Bobbies Hobbies
Generated two databases with one containing the names and ages of a group of people and the other containing the id and hobbies of the same group of people. Initially I added a new person and their hobby to the respective databases. My first query explicitly joined the two databases and showed the names and hobbies of the people in the group by checking if the primary key id in first database matched the person_id in the second database. My second query used an inner join which did the same as the first query but I limited the information shown by only including people whose name in the first database was Bobby McBobbyFace.