Create Customer orders
Generated two databases, the first was called customers and it contained the customers name and email while the second database contained the customers id, the item they purchased, and the price of the item. My first query used a left outer join to merge the customer and order databases where the id in the customer id in the customer database equaled the customer_id in the orders database and showed the customers name, email, item, and item price. I used a left outer join to show all null values since one of the customers had not made any purchases. My second query was more or less the same except that it showed the customers name, email, and total amount they spent on their orders. I grouped by the customers name since names were the most unique point of data.