Create Sequels in SQL
Generated a database that contained the title, release year, and sequel number for the Harry Potter films. The query used is a self join that displays the title of the Harry Potter movie next to the title of the movie that comes next in the series. A left outer join was used since the last movie does not have a sequel so it would return a NULL on a regular join.