Create Friendbook
Generated three databases where one contained the name and age of a group of people, the second contained the personal id related to the first table and the hobbies associated with them, and the third database contained personal id pairings that denoted who in the first database was friends with who. The first query displays the full names of the people in the first database and their associated hobbies from the second database. The second query used a multiple JOINs to display the names of the people who were friends with each other.