Create Dialects Practice
Accessed two databases containing information on Netflix shows and movies. The first query used a LEFT JOIN to combine and display the medias id, director, title, and type; a LEFT JOIN was used in case any of the selected columns were missing information. The second query showed how many movies and how many tv shows were listed in the database. The third query listed all of the media in the database in alphabetical order. The fourth query found and displayed the director for the movie Bright Star by referencing the different databases. The fifth query found the oldest movies in the databases based on their release year and displayed the titles.