SQLite Practice
This was to practice and familiarize myself with SQLite. Given a a database with multiple tables I counted how many sales occurred in the month of January, how many of those sales were iPhones, I found the account numbers of the people who made purchases in the month of February, of the products sold in January which was the cheapest and what was it's price, what the total revenue for each item sold in January was, information regrading products sold at a specific location, and how many customers ordered more than two products at a time in February along with the average amount spent by those customers.